% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ackley.R
\name{ackley}
\alias{ackley}
\title{Ackley}
\usage{
ackley(xx, a = 20, b = 0.2, c = 2 * pi)
}
\arguments{
\item{xx}{numeric vector}

\item{a}{constant (optional), with default value 20}

\item{b}{constant (optional), with default value 0.2}

\item{c}{constant (optional), with default value 2*pi}
}
\value{
value obtained from evaluating ackley function at xx
}
\description{
Ackley
}
\examples{
ackley(c(1,2))
ackley(runif(10,-32,32))
}
