% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branin.R
\name{branin}
\alias{branin}
\title{Branin}
\usage{
branin(xx)
}
\arguments{
\item{xx}{numeric vector}
}
\value{
value obtained from evaluating branin function at xx
}
\description{
Branin
}
\examples{
branin(c(0.5,0.5))
}
