% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egg_holder.R
\name{egg}
\alias{egg}
\title{Egg Holder}
\usage{
egg(xx)
}
\arguments{
\item{xx}{numeric vector}
}
\value{
value obtained from evaluating egg holder function at xx
}
\description{
Egg Holder
}
\examples{
egg(c(10,10))
}
