% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hartmann3D.R
\name{hart3}
\alias{hart3}
\title{Hartmann 3D}
\usage{
hart3(xx)
}
\arguments{
\item{xx}{numeric vector}
}
\value{
value obtained from evaluating Hartmann3D function at xx
}
\description{
Hartmann 3D
}
\examples{
hart3(c(0.5,0.2,0.3))
}
