% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hartmann4D.R
\name{hart4}
\alias{hart4}
\title{Hartmann 4D}
\usage{
hart4(xx)
}
\arguments{
\item{xx}{numeric vector}
}
\value{
value obtained from evaluating Hartmann4D function at xx
}
\description{
Hartmann 4D
}
\examples{
hart4(c(0.2,0.5,0.3,0.5))
}
