% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hartmann6D.R
\name{hart6}
\alias{hart6}
\title{Hartmann 6D}
\usage{
hart6(xx)
}
\arguments{
\item{xx}{numeric vector}
}
\value{
value obtained from evaluating Hartmann6D function at xx
}
\description{
Hartmann 6D
}
\examples{
hart6(c(0.2,0.5,0.3,0.5,0.8,0.2))
}
