% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levy.R
\name{levy}
\alias{levy}
\title{Levy}
\usage{
levy(xx)
}
\arguments{
\item{xx}{numeric vector}
}
\value{
value obtained from evaluating Levy function at xx
}
\description{
Levy
}
\examples{
levy(runif(10,-10,10))
}
