% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matya.R
\name{matya}
\alias{matya}
\title{Matya}
\usage{
matya(xx)
}
\arguments{
\item{xx}{numeric vector}
}
\value{
value obtained from evaluating Matya function at xx
}
\description{
Matya
}
\examples{
matya(runif(2,-10,10))
}
