% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rastrigin.R
\name{rastr}
\alias{rastr}
\title{Rastrigin}
\usage{
rastr(xx)
}
\arguments{
\item{xx}{numeric vector}
}
\value{
value obtained from evaluating Rastrigin function at xx
}
\description{
Rastrigin
}
\examples{
rastr(runif(10,-5,5))
}
