% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Schwef.R
\name{schwef}
\alias{schwef}
\title{Schwefel}
\usage{
schwef(xx)
}
\arguments{
\item{xx}{numeric vector}
}
\value{
value obtained from evaluating Schwefel function at xx
}
\description{
Schwefel
}
\examples{
schwef(runif(10,-400,300))
}
