% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trid.R
\name{trid}
\alias{trid}
\title{Trid}
\usage{
trid(xx)
}
\arguments{
\item{xx}{numeric vector}
}
\value{
value obtained from evaluating Trid function at xx
}
\description{
Trid
}
\examples{
d <- 10
trid(runif(d,-d^2,d^2))
}
