% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trid.R
\name{trid_min}
\alias{trid_min}
\title{Find Trid Global Minimum for d Dimensions}
\usage{
trid_min(d)
}
\arguments{
\item{d}{numeric value specifying the desired dimension}
}
\value{
Global minimum for Trid function at 'd' dimensions
}
\description{
Find Trid Global Minimum for d Dimensions
}
\examples{
trid_min(10)
}
