% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trid.R
\name{trid_xi}
\alias{trid_xi}
\title{Trid values for minimum}
\usage{
trid_xi(d)
}
\arguments{
\item{d}{numeric value specifying the desired dimension}
}
\value{
The corresponding input values to obtain global mimimum for Trid function at 'd' dimensions
}
\description{
Trid values for minimum
}
\examples{
trid_xi(10)
}
