% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zakharov.R
\name{zakharov}
\alias{zakharov}
\title{Zakharov}
\usage{
zakharov(xx)
}
\arguments{
\item{xx}{numeric vector}
}
\value{
value obtained from evaluating Zakharov function at xx
}
\description{
Zakharov
}
\examples{
zakharov(runif(10,-5,10))
}
